feat(osd-viewer-demo): add page for offscreen canvas
